requirements.txt
google-generativeai
python-frontmatter
PyYAML
python-dotenv
supabase
rdflib
PyGithub
GitPython>=3.1.30 # Specifica una versione minima per la compatibilità